refacto: ContactList onClick on ContactRow
